Yelp/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pymongo==4.7.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pillow==11.0.0
python-dotenv==1.0.1
tavily-python==0.5.0
langchain==1.2.15
langchain-groq==1.1.2
kafka-python==2.3.0
